+++ src/main/java/reciter/algorithm/feedback/article/scorer/ReciterFeedbackArticleScorer.java
@@ -276,10 +276,11 @@ public class ReciterFeedbackArticleScorer extends AbstractFeedbackArticleScorer 
 				mapConsolidatedCSVData(articleMap,csvPrinter,personIdentifier);
 
 				csvPrinter.flush();
+				NeuralNetworkModelArticlesScorer nnmodel = new NeuralNetworkModelArticlesScorer();																	  
 				log.warn("Uploading CSV into S3 starts here******************",outputStream.toString(StandardCharsets.UTF_8.name()),filePath.toString());
 				boolean uploadCsvToS3 = uploadCsvToS3(outputStream.toString(StandardCharsets.UTF_8.name()),filePath.toString());
 				if(uploadCsvToS3) {
-					 deleteFile(filePath);
+					 nnmodel. deleteFile(filePath);
 					 log.info("File deleted successfully: " + filePath);
 				}
 				log.warn("Uploading CSV into S3 ends here******************");
@@ -330,8 +331,9 @@ public class ReciterFeedbackArticleScorer extends AbstractFeedbackArticleScorer 
 				csvPrinter.flush();
 				log.warn("Uploading CSV into S3 starts here******************",outputStream.toString(StandardCharsets.UTF_8.name()),filePath.toString());
 				boolean uploadCsvToS3 = uploadCsvToS3(outputStream.toString(StandardCharsets.UTF_8.name()),filePath.toString());
+				NeuralNetworkModelArticlesScorer nnmodel = new NeuralNetworkModelArticlesScorer();
 				if(uploadCsvToS3) {
-					 deleteFile(filePath);
+					 nnmodel.deleteFile(filePath);
 					 log.info("File deleted successfully: " + filePath);
 				}
 				log.warn("Uploading CSV into S3 ends here******************");
@@ -351,6 +353,8 @@ public class ReciterFeedbackArticleScorer extends AbstractFeedbackArticleScorer 
 					mapItemLevelCSVData(reCiterArticles,csvPrinter,personIdentifier);
 				
 				csvPrinter.flush();
+				NeuralNetworkModelArticlesScorer nnmodel = new NeuralNetworkModelArticlesScorer();
+				nnmodel.deleteFile(filePath);																	  
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
@@ -536,6 +540,7 @@ public class ReciterFeedbackArticleScorer extends AbstractFeedbackArticleScorer 
 		boolean isS3UploadRequired = isS3UploadRequired();
 		String feedbackIdentityS3BucketName = getProperty("aws.s3.feedback.score.bucketName");
         try {
+			NeuralNetworkModelArticlesScorer nnmodel = new NeuralNetworkModelArticlesScorer();																			   
         	  if(isS3UploadRequired) 
         	  {
         		  File jsonFile = new File(fileName);
@@ -543,8 +548,12 @@ public class ReciterFeedbackArticleScorer extends AbstractFeedbackArticleScorer 
         		// Write the User object to the JSON file
                   objectMapper.writeValue(jsonFile, articleIdentityFeedbackScore);
                   log.info("JSON data written to file successfully: ", jsonFile.getAbsolutePath());
-                  uploadJsonFileIntoS3(fileName, jsonFile);
-
+                  boolean uploadJsonFileIntoS3 = uploadJsonFileIntoS3(fileName, jsonFile);
+                  
+                  if(uploadJsonFileIntoS3) {
+                	  nnmodel.deleteFile(jsonFile.toPath());
+ 					 log.info("File deleted successfully: " + jsonFile);
+ 				}
         	  }
         	  else
         	  {	  
@@ -553,7 +562,7 @@ public class ReciterFeedbackArticleScorer extends AbstractFeedbackArticleScorer 
 				  log.info("JSON written to file successfully.", jsonFile.getAbsolutePath() +"-" + fileName);
         	  }
         	  String isS3UploadRequiredString = Boolean.toString(isS3UploadRequired);
-			  NeuralNetworkModelArticlesScorer nnmodel = new NeuralNetworkModelArticlesScorer();
+			  
 			  JSONArray articlesIdentityFeedbackScoreTotal = nnmodel.executeArticleScorePredictor("FeedbackIdentityScore", "feedbackIdentityScoreArticles.py",fileName,feedbackIdentityS3BucketName,isS3UploadRequiredString);
 			  log.info("articlesIdentityFeedbaclScoreTotal length",articlesIdentityFeedbackScoreTotal!=null?articlesIdentityFeedbackScoreTotal.length():0);
 			  if(articlesIdentityFeedbackScoreTotal!=null && articlesIdentityFeedbackScoreTotal.length() > 0)
@@ -712,7 +721,7 @@ public class ReciterFeedbackArticleScorer extends AbstractFeedbackArticleScorer 
 	    	article.setAuthorshipLikelihoodScore(0.0);
 	    return article; // Return null if not found
 	}
-	private void uploadJsonFileIntoS3(String keyName,File file)
+	private boolean uploadJsonFileIntoS3(String keyName,File file)
 	{
 		String FeedbackScoreBucketName = getProperty("aws.s3.feedback.score.bucketName");
         
@@ -725,51 +734,38 @@ public class ReciterFeedbackArticleScorer extends AbstractFeedbackArticleScorer 
 					.withRegion(System.getenv("AWS_REGION"))
 					.build();
         	
-        	log.info("Uploading files to S3 bucket ",FeedbackScoreBucketName);
-        	PutObjectRequest putObjectRequest = new PutObjectRequest(FeedbackScoreBucketName.toLowerCase(), keyName, file);
-       
-        	// Optionally, set metadata
-            ObjectMetadata metadata = new ObjectMetadata();
-            metadata.setContentType("application/json");
-            putObjectRequest.setMetadata(metadata);
-
-            
-            try{
-				s3.putObject(putObjectRequest);
-			    log.info("CSV file uploaded successfully to S3 bucket: " + FeedbackScoreBucketName);
+			if(s3.doesBucketExistV2(FeedbackScoreBucketName)) 
+			{												
+	        	log.info("Uploading files to S3 bucket ",FeedbackScoreBucketName);
+	        	PutObjectRequest putObjectRequest = new PutObjectRequest(FeedbackScoreBucketName.toLowerCase(), keyName, file);
+	       
+	        	// Optionally, set metadata
+	            ObjectMetadata metadata = new ObjectMetadata();
+	            metadata.setContentType("application/json");
+	            putObjectRequest.setMetadata(metadata);
+	
+	            
+	            try{
+					s3.putObject(putObjectRequest);
+				    log.info("CSV file uploaded successfully to S3 bucket: " + FeedbackScoreBucketName);
+					return true;   
+				}
+				catch(AmazonServiceException e) {
+					// The call was transmitted successfully, but Amazon S3 couldn't process 
+		            // it, so it returned an error response.
+					log.error(e.getErrorMessage());
+					 return false;
+				}
 			}
-			catch(AmazonServiceException e) {
-				// The call was transmitted successfully, but Amazon S3 couldn't process 
-	            // it, so it returned an error response.
-				log.error(e.getErrorMessage());
-			}
+        	else {
+        		log.error("S3 bucket does not exist: " + FeedbackScoreBucketName);
+                return false;
+        	}
         
         } catch (Exception e) {
             e.printStackTrace();
+			return false;			 
         }
 	}
 
-	private void deleteFile(Path filePath) {
-		try {
-			File file = filePath.toFile();
-			if (file.exists()) {
-
-				if (!file.setWritable(true)) {
-					log.info("Failed to set write permission for the file."+filePath);
-				}
-
-				if (Files.deleteIfExists(filePath)) {
-					log.info("File deleted successfully."+filePath);
-				} else {
-					log.info("File deletion failed."+filePath);
-				}
-
-			} else {
-				log.info("File does not exist: " + filePath);
-			}
-		} catch (Exception e) {
-			log.error("An error occurred while deleting the file path " + filePath + ": " + e.getMessage());
-		}
-
-	}									 
 }
